requirements.txt
groq
pymupdf
python-docx
streamlit
pillow
